% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{sun-methods}
\alias{crepuscule}
\alias{crepuscule,SpatialPoints,POSIXct-method}
\alias{crepuscule,matrix,POSIXct-method}
\alias{crepuscule-methods}
\alias{solarnoon}
\alias{solarnoon,SpatialPoints,POSIXct-method}
\alias{solarnoon,matrix,POSIXct-method}
\alias{solarnoon-methods}
\alias{solarpos}
\alias{solarpos,SpatialPoints,POSIXct-method}
\alias{solarpos,matrix,POSIXct-method}
\alias{solarpos-methods}
\alias{sun-methods}
\alias{sunriset}
\alias{sunriset,SpatialPoints,POSIXct-method}
\alias{sunriset,matrix,POSIXct-method}
\alias{sunriset-methods}
\title{Methods for sun ephemerides calculations}
\arguments{
\item{crds}{a \code{SpatialPoints} or \code{matrix} object, containing x and
y coordinates (in that order).}

\item{dateTime}{a POSIXct object with the date and time associated to
calculate ephemerides for points given in crds.}

\item{solarDep}{numeric vector with the angle of the sun below the horizon
in degrees.}

\item{direction}{one of "dawn", "dusk", "sunrise", or "sunset", indicating
which ephemerides should be calculated.}

\item{POSIXct.out}{logical indicating whether POSIXct output should be
included.}

\item{proj4string}{string with valid projection string describing the
projection of data in \code{crds}.}

\item{\dots}{other arguments passed through.}
}
\value{
\code{crepuscule}, \code{sunriset}, and \code{solarnoon} return a
numeric vector with the time of day at which the event occurs, expressed as
a fraction, if POSIXct.out is FALSE; otherwise they return a data frame with
both the fraction and the corresponding POSIXct date and time.
\code{solarpos} returns a matrix with the solar azimuth (in degrees from
North), and elevation.
}
\description{
Functions for calculating sunrise, sunset, and times of dawn and dusk, with
flexibility for the various formal definitions.  They use algorithms
provided by the National Oceanic & Atmospheric Administration (NOAA).
}
\details{
NOAA used the reference below to develop their Sunrise/Sunset

\url{http://www.srrb.noaa.gov/highlights/sunrise/sunrise.html} and Solar
Position

\url{http://www.srrb.noaa.gov/highlights/sunrise/azel.html} Calculators.
The algorithms include corrections for atmospheric refraction effects.

Input can consist of one location and at least one POSIXct times, or one
POSIXct time and at least one location.  \var{solarDep} is recycled as
needed.

Do not use the daylight savings time zone string for supplying
\var{dateTime}, as many OS will not be able to properly set it to standard
time when needed.
}
\note{
NOAA notes that \dQuote{for latitudes greater than 72 degrees N and S,
calculations are accurate to within 10 minutes.  For latitudes less than +/-
72 degrees accuracy is approximately one minute.}
}
\section{Warning}{
 Compared to NOAA's original Javascript code, the sunrise
and sunset estimates from this translation may differ by +/- 1 minute, based
on tests using selected locations spanning the globe. This translation does
not include calculation of prior or next sunrises/sunsets for locations
above the Arctic Circle or below the Antarctic Circle.
}
\author{
Sebastian P. Luque \email{spluque@gmail.com}, translated from Greg
Pelletier's \email{gpel461@ecy.wa.gov} VBA code (available from
\url{http://www.ecy.wa.gov/programs/eap/models.html}), who in turn
translated it from original Javascript code by NOAA (see Details).  Roger
Bivand \email{roger.bivand@nhh.no} adapted the code to work with \pkg{sp}
classes.
}
\references{
Meeus, J.  (1991) Astronomical Algorithms.  Willmann-Bell, Inc.
}
\keyword{manip}
\keyword{methods}
\keyword{utilities}

